% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWLelast.cw.bw.R
\name{GWLelast.cv.bw}
\alias{GWLelast.cv.bw}
\title{GWLelast.cv.bw}
\usage{
GWLelast.cv.bw(x = x, y = y, D = D, coords = coords, alpha = 1,
  lambda = lambda, nlambda = nlambda, gweight = gweight,
  longlat = longlat, bw = bw)
}
\arguments{
\item{x}{Covariates.}

\item{y}{Outcome binary variable.}

\item{D}{Distance matrix.}

\item{coords}{2 columns matrix including "longitude" and "latitude".}

\item{alpha}{The elasticnet mixing parameter [0,1] in glmnet package.}

\item{lambda}{Optional user-supplied lambda sequence in glmnet package.}

\item{nlambda}{The number of lambda values in glmnet package.}

\item{gweight}{geographical kernel function in spgwr package.}

\item{longlat}{Indicate if the coords parameter are sperically calculated.}

\item{bw}{bandwidth of geographical kernel function.}
}
\value{
error Cross validation error.
}
\description{
Cross validation for geographically weighted logistic elastic net regression
}
