% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GWLelast.inner.parallel.R
\name{GWLelast.inner.parallel}
\alias{GWLelast.inner.parallel}
\title{Inner part of fitting GWLelast with parallel cores}
\usage{
GWLelast.inner.parallel(x = x, y = y, coords = coords, W = W,
  alpha = 1, lambda = lambda, nlambda = nlambda)
}
\arguments{
\item{x}{Covariates.}

\item{y}{Outcome binary variable.}

\item{coords}{2 columns matrix including "longitude" and "latitude".}

\item{W}{Weight matrix.}

\item{alpha}{The elasticnet mixing parameter [0,1] in glmnet package.}

\item{lambda}{Optional user-supplied lambda sequence in glmnet package.}

\item{nlambda}{The number of lambda values in glmnet package.}
}
\value{
model Fitted model at location i.

cv.error Cross validation error.
}
\description{
Inner part of fitting GWLelast with parallel cores
}

