% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWLelast.R
\docType{package}
\name{GWLelast}
\alias{GWLelast}
\alias{GWLelast-package}
\title{GWLelast
Geographically weighted logistic elastic net regression}
\description{
This pacakge fits the geographically weighted logistic elastic net regression model for a valible seelction and for the mitigatin of the multicolinearity between coefficients due to geographical correlation. Detailed explanations can be found in Yoneoka et al. (2016): New algorithm for constructing area-based index with geographical heterogeneities and variable selection: An application to gastric cancer screening.
}
