% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWLelast.sel.bw.R
\name{GWLelast.sel.bw}
\alias{GWLelast.sel.bw}
\title{GWLelast.sel.bw}
\usage{
GWLelast.sel.bw(x, y, coords, D = NULL, alpha = 1, lambda = NULL,
  nlambda = NULL, gweight = gweight, longlat = TRUE, lower.bw = NULL,
  upper.bw = NULL)
}
\arguments{
\item{x}{Covariates.}

\item{y}{Outcome binary variable.}

\item{coords}{2 columns matrix including "longitude" and "latitude".}

\item{D}{Distance matrix.}

\item{alpha}{The elasticnet mixing parameter [0,1] in glmnet package.}

\item{lambda}{Optional user-supplied lambda sequence in glmnet package.}

\item{nlambda}{The number of lambda values in glmnet package.}

\item{gweight}{geographical kernel function in spgwr package.}

\item{longlat}{Indicate if the coords parameter are sperically calculated.}

\item{lower.bw}{Lower limit of bandwidth in geographical kernel.}

\item{upper.bw}{Upper limit of bandwidth in geographical kernel.}
}
\value{
optimal.bw Optimal bandwidth.
}
\description{
Bandwidth selection forgeographically weighted logistic elastic net regression
}
\examples{
######################
# Need to add
}
