\name{GWLelastic-package}
\alias{GWLelastic-package}
\alias{GWLelastic}
\docType{package}
\title{
GWLelast
}
\description{
This pacakge fits the geographically weighted logistic elastic net regression model for a valible seelction and for the mitigatin of the multicolinearity between coefficients due to geographical correlation.

This is an extention of geographically weighted lasso proposed by Wheeler 2009.

}
\details{
\tabular{ll}{
Package: \tab GWLelastic\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-26\cr
License: \tab GPL-2\cr
}
}
\author{
Daisuke Yoneoka <blue.sky.sea.dy@gmail.com>, Eiko Saito
}
\references{
Wheeler, D.C. (2009). Simultaneous coefficient penalization and model selection in geographically weighted regression: the geographically weighted lasso. Environment and Planning A 41, 7220742
}
\keyword{package}