% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GWLelast.est.R
\name{GWLelast.est}
\alias{GWLelast.est}
\title{Fitting geographically weighted logistic elastic net regression}
\usage{
GWLelast.est(x, y, coords, alpha = 1, lambda = NULL, nlambda = NULL,
  gweight = c("gwr.Gauss", "gwr.bisquare"), longlat = TRUE, bw = bw,
  D = NULL, Parallel = FALSE)
}
\arguments{
\item{x}{Covariates.}

\item{y}{Outcome binary variable.}

\item{coords}{2 columns matrix including "longitude" and "latitude".}

\item{alpha}{The elasticnet mixing parameter [0,1] in glmnet package.}

\item{lambda}{Optional user-supplied lambda sequence in glmnet package.}

\item{nlambda}{The number of lambda values in glmnet package.}

\item{gweight}{geographical kernel function in spgwr package.}

\item{longlat}{Indicate if the coords parameter are sperically calculated.}

\item{bw}{bandwidth of geographical kernel function.}

\item{D}{Distance matrix.}

\item{Parallel}{Calculate the model with multi core or not.}
}
\value{
model Fitted model at location i.

cv.error Cross validation error.
}
\description{
Fitting geographically weighted logistic elastic net regression
}

