% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GWLelastic.R
\docType{package}
\name{GWLelast}
\alias{GWLelast}
\alias{GWLelast-package}
\title{Geographically weighted logistic elastic net regression}
\description{
This pacakge fits the geographically weighted logistic elastic net regression model for a valible seelction and for the mitigatin of the multicolinearity between coefficients due to geographical correlation.
}

