% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GWLelast.sel.bw.R
\name{GWLelast.sel.bw}
\alias{GWLelast.sel.bw}
\title{Bandwidth selection forgeographically weighted logistic elastic net regression}
\usage{
GWLelast.sel.bw(x, y, coords, alpha = 1, lambda = NULL, nlambda = NULL,
  gweight = gweight, longlat = TRUE, lower.bw = NULL, upper.bw = NULL,
  D = NULL, tol = .Machine$double.eps^0.25, Parallel = FALSE)
}
\arguments{
\item{x}{Covariates.}

\item{y}{Outcome binary variable.}

\item{coords}{2 columns matrix including "longitude" and "latitude".}

\item{alpha}{The elasticnet mixing parameter [0,1] in glmnet package.}

\item{lambda}{Optional user-supplied lambda sequence in glmnet package.}

\item{nlambda}{The number of lambda values in glmnet package.}

\item{gweight}{geographical kernel function in spgwr package.}

\item{longlat}{Indicate if the coords parameter are sperically calculated.}

\item{lower.bw}{Lower limit of bandwidth in geographical kernel.}

\item{upper.bw}{Upper limit of bandwidth in geographical kernel.}

\item{D}{Distance matrix.}

\item{tol}{The desired accuracy in optimize function.}

\item{Parallel}{Calculate the model with multi core or not.}
}
\value{
optimal.bw Optimal bandwidth.
}
\description{
Bandwidth selection forgeographically weighted logistic elastic net regression
}

